# seaf-archtool-core/public/documentation/docs/manual/backend/api.yaml
...
openapi: 3.0.3
info:
  version: 1.0.0
  title: SEAF.ArchTool API
  description: >
    Данный документ актуален для развертывания в режимах Client-Server и Cluster.
    Здесь предоставляются интеграционные API, которые вы можете использовать для 
    встраивания SEAF.ArchTool в свои производственные процессы.
    Подробнее тут https://gitverse.ru/seafteam/seaf-archtool-core/content/master/public/documentation/docs/manual/config/deployment.md
servers:
  - url: https://{URL}
    description: API server secure
    variables:
      URL:
        default: localhost:3030
        description: Сервер https для запроса
  - url: http://{URL}
    description: API server insecure
    variables:
      URL:
        default: localhost:3030
        description: Сервер http для запроса
...
paths:
  /seaf-core/api/core/storage/jsonata/{query}:
    get:
      tags:
      - DataLake
      summary: Выполняет произвольный запрос JSONata к данным архитектуры 
      parameters:
        - in: path
          name: query
          description: JSONata запрос в скобках "([запрос])" или идентификатор DataSet без скобок
          required: true
          schema:
            $ref: '#/components/schemas/JSONataQuery'
          example: >
            ({"id": $params.id, "limit": $params.limit, "type": $self.type, "template": $self.template})
        - in: query
          name: params
          description: > 
            JSON структура с параметрами передающаяся в запрос.
            Будет доступна в переменной $params.
          required: false
          schema:
            $ref: '#/components/schemas/JSONataParams'
          example: '{"id":"sber.rdwsoft","limit":8}'
        - in: query
          name: subject
          description: > 
            JSON структура с профилем объекта, который выполняет запрос.
            Будет доступна в переменной $self.
          required: false
          schema:
            $ref: '#/components/schemas/JSONataParams'
          example: '{"type":"markdown","template":"template.md"}'
      responses:
        "200":
          description: "Результат выполнения запроса"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONataResult'
  /seaf-core/api/core/storage/reload:
    put:
      tags:
        - DataLake
      summary: Перезагрузка манифестов в кластере
      description: |
        **В режиме backend**:
          - перезагружает манифест, дополнительные параметры не нужны и не учитываются в обработке, необходимо передать только secret
        
        **В режиме cluster**:
          - работает по указанным ниже параметрам. Умеет:
            - перезагружать выбранные манифесты
            - перезагружать все манифесты
            - обнаруживать ссылки на новые манифесты в root манифесте
        Перезагружает указанные манифесты или все манифесты в кластере.
        Может использоваться для принудительного обновления кэша.
        
        В случае возникновения ошибки, можно поискать в логах по значению 'trace-id-123' из ответа с ошибкой
      operationId: reloadClusterStorage
      parameters:
        - name: secret
          in: query
          description: secret для доступа к запросу. Указывается в .env переменной VUE_APP_DOCHUB_RELOAD_SECRET
          required: true
          schema:
            type: string
          example: "your-reload-secret-key"

        - name: reloadRoot
          in: query
          description: |
            Если true - запускает перезагрузку манифестов для подтягивания изменений из root манифеста. 
            Перезагрузка может запуститься если ранее она уже была запрошена (другим запросом) и не успела обработаться
            При этом параметры alias и mode игнорируются если переданы.
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
          example: "false"

        - name: mode
          in: query
          description: |
            Режим работы с параметром alias.
            - `alias` (по умолчанию) - alias трактуются как имена манифестов
            - `permission` - alias трактуются как разрешения, которые конвертируются в имена манифестов через root manifest
          schema:
            type: string
            enum: [alias, permission]
            default: alias
          example: "alias"

        - name: alias
          in: query
          description: |
            Имена манифестов для перезагрузки.
            
            **Формат передачи:**
            - Для одного значения: `?alias=manifest1`
            - Для нескольких значений: `?alias=manifest1&alias=manifest2&alias=manifest3`
            
            Если параметр не указан - перезагружаются все доступные манифесты.
            
            **Важно:** При `mode=permission` этот параметр содержит имена разрешений,
            а не имена манифестов.
            
            **Формат при `mode=permission` не отличается, но в alias ожидаются значения permission :**
            - Для одного значения: `?alias=permission1`
            - Для нескольких значений: `?alias=permission1&alias=permission2&alias=permission3`
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: ["manifest1"]

      responses:
        '200':
          description: Запрос успешно обработан
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                example: {
                  "message": "success"
                }

        '400':
          description: |
            Неверные параметры запроса
            В режиме backend если что-то пошло не так, тоже возвращается 400
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: bad params 'alias' (nir)"

        '403':
          description: Неверный секретный ключ
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: Error reload secret is not valid [wrong-secret]"

        '500':
          description: Внутренняя ошибка сервера или сервер не готов
          headers:
            Retry-After:
              description: Время в секундах, через которое можно повторить запрос
              schema:
                type: string
                example: "60"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: server not ready (aliases empty in storage)"
  /seaf-core/api/core/storage/release-data-profile/{path}:
    get:
      tags:
      - DataLake
      summary: >
        По указанному пути возвращает результат реализации профиля данных.
        Подробнее здесь - https://gitverse.ru/seafteam/seaf-archtool-core/content/master/public/documentation/docs/manual/datasets.md
      parameters:
        - in: path
          name: path
          description: Путь к профилю данных
          required: true
          schema:
            $ref: '#/components/schemas/PathToDataProfile'
          example: "/entities/interactions/presentations/blank"
        - in: query
          name: params
          description: > 
            JSON структура с параметрами передающаяся в запрос.
            Будет доступна в переменной $params.
          required: false
          schema:
            $ref: '#/components/schemas/JSONataParams'
          example: '{"id":"archtool.user"}'
      responses:
        "200":
          description: "Результат выполнения запроса"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONataResult'
...
  /seaf-core/api/core/storage/put-content/{hash}:
    post:
      tags:
        - DataLake
      summary: Сохраняет данные в Bitbucket
      parameters:
        - in: path
          name: hash
          description: Хеш пути к сущности
          required: true
          schema:
            type: string
          example: 8743b52063cd84097a65d1633f5c74f5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
                description: В теле запроса передается произвольное количество пар ключ-значение, где ключ - путь к сущности, значение - контент в формате JSON строки
            example:
              file-name.yaml: "updated content is very beautiful"
              file-name2.yaml: "updated content is very beautiful here too"
      responses:
        "200":
          description: "Данные обновлены успешно"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Created"
        "404":
          description: "В конфигурации указана СКВ отличная от Bitbucket"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Данное API поддерживает только Bitbucket Open API"
        "400":
          description: "Получено пустое тело запроса"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Отсутствует тело запроса"
        "500":
          description: "Внутренняя ошибка сервера. Данная ошибка возникает если не удалось извлечь тело запроса или Bitbucket вернул ошибку"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Получено пустое тело запроса"
...
  /seaf-core/api/core/user-info:
    get:
      tags:
        - UserInfo
      summary: Получает информацию о текущем пользователе
      description: Данные о пользователе получаются из IdP токена
      security:
        - idpAuth: [ ]
      responses:
        "200":
          description: Успешный ответ с данными пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  surname:
                    type: string
                    example: Иванов
                  name:
                    type: string
                    example: Иван
                  roles:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                          example: ADMIN
                        label:
                          type: string
                          example: Администратор
                  initials:
                    type: string
                    example: ИИ
                  userId:
                    type: string
                    example: 11879000
...
components:
  schemas:
    JSONataQuery:
      type: string
      pattern: '^([a-zA-Z\_]*(\.[a-zA-Z\_][a-zA-Z\_]*)*)|(\(.*\))$'
    JSONataParams:
      type: string
    JSONataResult:
      type: object
    PathToDataProfile:
      type: string
      pattern: '^\/[a-zA-Z][a-zA-Z\_0-9]*(\/[a-zA-Z][a-zA-Z\_0-9]*)*$'
...